met_annot_unifier/aligner/parser.py
import pandas as pd


def parse_gnps(file_path):
    # Read the GNPS output file
    data = pd.read_csv(file_path, sep="\t")
    # Extract necessary columns and any other processing
    return data


def parse_sirius(file_path):
    # Read the Sirius output file
    data = pd.read_csv(file_path, sep="\t")
    # Extract necessary columns and any other processing
    return data


def parse_isdb(file_path):
    # Read the ISDB output file
    data = pd.read_csv(file_path, sep="\t")
    # Extract necessary columns and any other processing
    return data


def standardize_column_names(df, original_column_name, final_column_name):
    """
    Standardizes the column names of a dataframe.

    Args:
    df (pandas.DataFrame): The dataframe to standardize.
    original_column_name (str): The current name of the column to be standardized.
    final_column_name (str): The final standardized name for the column.

    Returns:
    pandas.DataFrame: The dataframe with standardized column names.

    Example:
    >>> df = pd.DataFrame({'InChIKey-Planar': [1, 2], 'OtherColumn': [3, 4]})
    >>> standardized_df = standardize_column_names(df, 'InChIKey-Planar', 'InChiKey')
    >>> standardized_df.columns
    Index(['InChiKey', 'OtherColumn'], dtype='object')
    """
    standardized_df = df.rename(columns={original_column_name: final_column_name})
    return standardized_df
